+++ app/models/reports/net_suite/donation_export.rb
@@ -1,0 +1,10 @@
+require "csv"
+
+module Reports
+  module NetSuite
+    class DonationExport
+      def to_csv
+      end
+    end
+  end
+end
