+++ app/models/reports/net_suite/donor_export.rb
@@ -1,0 +1,10 @@
+require "csv"
+
+module Reports
+  module NetSuite
+    class DonorExport
+      def to_csv
+      end
+    end
+  end
+end
