+++ app/views/common/_navbar.html.erb
@@ -32,6 +32,8 @@
                 <li role="separator" class="divider"></li>
                 <li><%= link_to "Export Master Inventory", exports_path %></li>
                 <li role="separator" class="divider"></li>
+                <li><%= link_to "NetSuite Donations Export", net_suite_donation_export_reports_path %></li>
+                <li><%= link_to "NetSuite Donors Export", net_suite_donor_export_reports_path %></li>
                 <li><%= link_to "NetSuite Orders Export", net_suite_order_export_reports_path %></li>
               <% end %>
             </ul>
