+++ config/routes.rb
@@ -55,6 +55,8 @@ Rails.application.routes.draw do
   resources :reports, only: [] do
     collection do
       get :graphs
+      get :net_suite_donation_export
+      get :net_suite_donor_export
       get :net_suite_order_export
       get :total_inventory_value
       get :value_by_county
